% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{add_column_table}
\alias{add_column_table}
\title{Add table horizontally adjacent to plot}
\usage{
add_column_table(
  p,
  ...,
  left_cols = NULL,
  right_cols = NULL,
  left_args = NULL,
  right_args = NULL,
  widths = NULL,
  plot_margin = ggplot2::margin(t = 0, b = 0, r = 1, l = 1, unit = "pt")
)
}
\arguments{
\item{p}{ggplot object}

\item{...}{Arguments common to both left and right plot tables}

\item{right_cols, left_cols}{Character vector of column names in \code{p@data} to display in plot}

\item{left_args, right_args}{List of arguments used to generate plots to left and right of \code{p}}

\item{widths}{Widths of plots. Length must match \code{length(left_cols) + length(right_cols) + 1}. Default is \code{NULL}}

\item{plot_margin}{Margin for \code{p} and plots to right and left of \code{p}}
}
\value{
patchwork object
}
\description{
Add table horizontally adjacent to plot
}
