% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae}
\alias{ae}
\title{Vaccine safety (adverse event) data}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
ae
}
\description{
1 row for each combination of patient population, vaccine product, and AE outcome per study
}
\keyword{datasets}
