% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_friedman.R
\name{bin}
\alias{bin}
\title{Bin a numeric vector}
\usage{
bin(x, n_bins)
}
\arguments{
\item{x}{A numeric vector.}

\item{Integer}{specifying the number of bins to split \code{x} into.}
}
\description{
Function to bin a numeric vector
}
\keyword{internal}
