% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coadmin}
\alias{coadmin}
\title{Co-administration data}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
coadmin
}
\description{
Co-administration data
}
\keyword{datasets}
