% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comments}
\alias{comments}
\title{Comments in REDCap forms}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
comments
}
\description{
Comments in REDCap forms
}
\keyword{datasets}
