% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core}
\alias{core}
\title{Raw data from RedCAP}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
core
}
\description{
See df_core.R file in data-raw folder for reproducible workflow
1 row per study
}
\keyword{datasets}
