% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{crosstableServer}
\alias{crosstableServer}
\title{Module server for fluid page with cross table plot in main panel and plot settings in a sidebar}
\usage{
crosstableServer(id, data, plotly_toolbar_buttons = "toImage")
}
\arguments{
\item{id}{Input id. Must match \code{id} entered in \code{crosstableUI}}

\item{data}{Data frame for plotting}

\item{plotly_toolbar_buttons}{Buttons to include in plotly toolbar}
}
\value{
Reactive list containing "plot", "population", "virus", "vax_product", "x_var", "y_var"
}
\description{
Module server for fluid page with cross table plot in main panel and plot settings in a sidebar
}
