% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module_other.R
\name{crosstableUI}
\alias{crosstableUI}
\title{UI for side bar with cross table plot in main panel and plot settings in a sidebar}
\usage{
crosstableUI(
  id,
  sidebar_open = FALSE,
  sidebar_width = "30\%",
  plot_resizable = TRUE,
  plot_width = "100\%",
  plot_height = "400px",
  show_color_picker = FALSE,
  show_legend_switch = FALSE,
  primary_color = "#246A87",
  secondary_color = "#A63B86",
  accordion_btn_background_color_alpha = 0.1,
  accordion_btn_border_color = "white",
  slider_color = primary_color,
  switch_color = primary_color,
  incrementor_button_color = primary_color,
  color_by_virus = TRUE
)
}
\arguments{
\item{id}{Input id. Must match \code{id} entered in \code{crosstableServer}}

\item{sidebar_open}{If \code{TRUE} (default), sidebar is open by default}

\item{sidebar_width}{Width of sidebar. Default is \code{"30\%"}}

\item{plot_resizable}{If \code{TRUE} (default), plot can be resized by draggling bottom right corner}

\item{plot_width, plot_height}{Default width and height of plot. Default is \code{"100\%"} for \code{plot_width} and \code{"200px"} for \code{plot_height}}

\item{show_color_picker}{If \code{TRUE}, color picker is displayed above plot. Default is \code{FALSE}}

\item{show_legend_switch}{If \code{TRUE}, switch input is displayed to control whether plot legend is displayed. Default is \code{FALSE}}

\item{primary_color, secondary_color}{Primary and secondary color used for bootstrap theme. Default is \code{"#A63B86"} for \code{secondary_color}}

\item{accordion_btn_background_color_alpha}{Alpha filter for accordion buttons. Default is \code{0.1}}

\item{accordion_btn_border_color}{Color for accordion buttons}

\item{slider_color, switch_color}{Color for sliders and switches}

\item{incrementor_button_color}{Color for incrementor button}

\item{color_by_virus}{If \code{TRUE} (default), \code{plot_crosstable2} is used as default plot function in server. If \code{FALSE}, \code{plot_crosstable} is used as default plot function in server}
}
\value{
Enter as input to UI
}
\description{
UI for side bar with cross table plot in main panel and plot settings in a sidebar
}
