% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_firm.R
\name{vi_pdp}
\alias{vi_pdp}
\alias{vi_ice}
\title{Deprecated}
\usage{
vi_pdp(...)

vi_ice(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link{vi_firm}}.}
}
\description{
These functions have been deprecated and should not be used. They will be
removed in the next release.
}
