% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_shiny}
\alias{df_shiny}
\title{Data for 1st version of shiny app}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
df_shiny
}
\description{
Data for 1st version of shiny app
}
\keyword{datasets}
