% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Data dictionary}
\format{
Data frame
\describe{
\item{original}{Variable name in csv file directly exported from RedCAP}
\item{new}{Variable name in \code{core}}
\item{description}{Variable definition used in RedCAP entry form}
}
}
\usage{
dict
}
\description{
Data dictionary
}
\keyword{datasets}
