% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain}
\alias{domain}
\title{Domain data for all studies}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
domain
}
\description{
Domain data for all studies
}
\keyword{datasets}
