% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{editTableServer}
\alias{editTableServer}
\title{Module server for data table}
\usage{
editTableServer(
  id,
  df,
  allowRowEdit = FALSE,
  allowColumnEdit = FALSE,
  manualRowMove = TRUE,
  ...
)
}
\arguments{
\item{id}{Input id. Must match \code{id} entered in \code{editTableUI}}

\item{df}{Data frame for plotting}

\item{allowRowEdit}{If \code{TRUE}, data in table can be edited within app}

\item{allowColumnEdit}{If \code{TRUE}, column names in table can be edited within app}

\item{manualRowMove}{If \code{TRUE}, rows in table many be reordered within app}

\item{...}{Not used}
}
\value{
Enter into server of shiny app
}
\description{
Module server for data table
}
