% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{editTableUI}
\alias{editTableUI}
\title{Module UI for data table}
\usage{
editTableUI(id, ...)
}
\arguments{
\item{id}{Input id. Must match \code{id} entered in \code{editTableServer}}

\item{...}{Arguments passed to \code{rhandsontable::rHandsontableOutput}}
}
\value{
Enter as input to UI of shiny app
}
\description{
Module UI for data table
}
