% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epi}
\alias{epi}
\title{Epidemiological data for SARS-CoV2, RSV, and Influenza}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
epi
}
\description{
1 row for each combination of patient population and virus per study
}
\keyword{datasets}
