% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_control_matrix.R
\name{fill_2x2}
\alias{fill_2x2}
\title{Complete a partially filled 2 x 2 cross tabulation}
\usage{
fill_2x2(
  df,
  r1c1 = "n_vaxed_with_outcome",
  r2c1 = "n_unvaxed_with_outcome",
  r1c2 = "n_vaxed_without_outcome",
  r2c2 = "n_unvaxed_without_outcome",
  r1 = "n_vaxed",
  r2 = "n_unvaxed",
  c1 = "n_with_outcome",
  c2 = "n_without_outcome",
  total = "n_total",
  replace_columns = FALSE,
  haldane_correction = TRUE
)
}
\arguments{
\item{df}{Data frame containing raw counts in separate columns. Each row represents a single 2 x 2 matrix}

\item{r1c1, r1c2, r2c1, r2c2}{Either name of column in \code{df} (as length 1 character vector) containing raw counts for cell in 2 x 2 matrix or a vector of counts with length equal to \code{nrow(df)} (or 1). Use \code{NULL} if column not present in \code{df}}

\item{r1, r2, c1, c2}{Either name of column in \code{df} (as length 1 character vector) containing row/column sums in 2 x 2 matrix or a vector of counts with length equal to \code{nrow(df)} (or 1). Use \code{NULL} if column not present in \code{df}}

\item{total}{Either name of column in \code{df} (as length 1 character vector) containing grand total of cell counts in 2 x 2 matrix or a vector of counts with length equal to \code{nrow(df)} (or 1). Use \code{NULL} if column not present in \code{df}}

\item{replace_columns}{If \code{TRUE}, original count columns are replaced by calculated counts with new columns added for counts not included in \code{df}. If \code{FALSE} (default), calculated counts added as separate columns. Names for any newly added columns include: ".r1c1", ".r1c2", ".r2c1", ".r2c2", ".r1", ".r2", ".c1", ".c2", ".total". If naming collisions occur, prior column names will be overwritten with new column names}
}
\value{
Data frame with containing calculated counts. Regardless of input to \code{replace_columns}, output will include a logical column named ".calculable" indicating whether all counts could be calculated for a given row
}
\description{
Complete a partially filled 2 x 2 cross tabulation
}
