% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\docType{data}
\name{GeomStripes}
\alias{GeomStripes}
\title{Geom for alternating background zebra stripes}
\format{
An object of class \code{GeomStripes} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomStripes
}
\description{
Geom for alternating background zebra stripes
}
\keyword{datasets}
