% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_default_metric.ppr}
\alias{get_default_metric.ppr}
\title{#' @keywords internal
get_default_metric.lm <- function(object) {
  if (object$family$family == "binomial") {
    "auc"
  } else {
    "rmse"
  }
}}
\usage{
\method{get_default_metric}{ppr}(object)
}
\description{
#' @keywords internal
get_default_metric.lm <- function(object) {
  if (object$family$family == "binomial") {
    "auc"
  } else {
    "rmse"
  }
}
}
\keyword{internal}
