% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pd_vi_scores.R
\name{get_pd_vi_scores}
\alias{get_pd_vi_scores}
\title{Variable Importance}
\usage{
get_pd_vi_scores(object, pred.var, FUN = NULL, keep.partial = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{pred.var}{Character string giving the names of the predictor variables
of interest.}

\item{FUN}{Function used to measure the variability of the partial dependence
values for continuous predictors. If \code{NULL}, the standard deviation is
used (i.e., \code{FUN = sd}). For factors, the range statistic is used (i.e.,
(max - min) / 4).}

\item{keep.partial}{Logical indicating whether or not to return the computed
partial dependence values for each predictor listed in \code{pred.var} in a
separate attribute called \code{"partial"}. Default is \code{FALSE}.}

\item{...}{Additional optional arguments to be passed onto
\code{\link{get_pd_vi_score}}.}
}
\description{
Compute variable importance for a set of predictors using the partial
dependence-based approach.
}
\keyword{internal}
