% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred_names.R
\name{get_pred_names}
\alias{get_pred_names}
\title{Extract Predictor Names}
\usage{
get_pred_names(object, ...)
}
\arguments{
\item{object}{An appropriate fitted model object.}

\item{...}{Additional optional arguments.}
}
\description{
Extract predictor names from a fitted model.
}
\keyword{internal}
