% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{get_training_data.default}
\alias{get_training_data.default}
\title{#' @keywords internal
get_training_data.default <- function(object) {
  if (isS4(object)) {
    stop(msg)
  } else {
    train <- eval(stats::getCall(object)$data)
    if (is.null(train)) {
      stop(msg)
    } else {
      if (!(is.data.frame(train))) {
        if (is.matrix(train) || is.list(train)) {
          train <- as.data.frame(train)
        } else {
          stop(msg)
        }
      }
    }
  }
  train
}}
\usage{
\method{get_training_data}{default}(object, env = parent.frame(), arg = "data")
}
\description{
#' @keywords internal
get_training_data.default <- function(object) {
  if (isS4(object)) {
    stop(msg)
  } else {
    train <- eval(stats::getCall(object)$data)
    if (is.null(train)) {
      stop(msg)
    } else {
      if (!(is.data.frame(train))) {
        if (is.matrix(train) || is.list(train)) {
          train <- as.data.frame(train)
        } else {
          stop(msg)
        }
      }
    }
  }
  train
}
}
\keyword{internal}
