% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{updateIncrementor}
\alias{updateIncrementor}
\title{Update value of incrementor button}
\usage{
updateIncrementor(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL
)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{Input id for \code{incrementorInput}}

\item{value}{Updated value. Enter as length 1 numeric}
}
\value{
UI updated. Enter in server function
}
\description{
Update value of incrementor button
}
