% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{list_metrics}
\alias{list_metrics}
\title{List metrics}
\usage{
list_metrics()
}
\description{
List all available performance metrics.
}
\examples{
(metrics <- list_metrics())
metrics[metrics$Task == "Multiclass classification", ]
}
