% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module_meta.R
\name{metaAnalysisServer}
\alias{metaAnalysisServer}
\title{Server for meta-analysis module}
\usage{
metaAnalysisServer(
  id,
  raw_data,
  meta,
  plotly_toolbar_buttons = "toImage",
  id_export_plot_btn = "export_plot_btn"
)
}
\arguments{
\item{id}{Input id. Must match \code{id} entered in \code{crosstableUI}}

\item{raw_data, meta}{Data frames for raw data and final results of meta-analysis respectively}

\item{plotly_toolbar_buttons}{Buttons to include in plotly toolbar}

\item{id_export_plot_btn}{ID for button to export plot. Default is \code{"export_plot_btn"}}
}
\value{
Enter inside server function of shiny app
}
\description{
Server for meta-analysis module
}
