% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%T>\%}
\alias{\%T>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%T>\% rhs

lhs \%>\% rhs
}
\description{
See \code{\link[magrittr]{\%T>\%}} for more details.

See \code{\link[magrittr]{\%>\%}} for more details.
}
\keyword{internal}
