% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crosstable.R
\name{plot_crosstable}
\alias{plot_crosstable}
\title{Plot cross table displaying the number of studies in a given domain for every virus (or vaccine product)}
\usage{
plot_crosstable(
  df,
  x = "virus",
  y = "domain",
  color_var = "n",
  color_min = "#FFFFFF",
  color_max = "#A63B86",
  facet = NULL,
  na.rm = TRUE,
  show_missing_x = FALSE,
  show_missing_y = FALSE,
  show_empty_levels = TRUE,
  show_zeros = FALSE,
  show_legend = FALSE,
  legend_title = "No. of studies",
  x_axis_pos = "top",
  x_axis_label_angle = 45,
  x_axis_title = NULL,
  y_axis_title = NULL,
  font_size = 18,
  plot_margin = NULL,
  aspect_ratio = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{x, y}{Categorical variables that will form x and y axis, respectively. Enter as length 1 character vectors}

\item{color_var}{Variable used to determine fill color of table. Default is \code{"n"}. Enter as length 1 character vectors}

\item{color_min, color_max}{Minimum and maximum values for continuous color scale. Enter as length 1 character vectors of hexadecimal codes or color names}
}
\value{
ggplot object
}
\description{
Plot cross table displaying the number of studies in a given domain for every virus (or vaccine product)
}
