% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crosstable.R
\name{plot_crosstable2}
\alias{plot_crosstable2}
\title{Version of \code{plot_crosstable} that uses separate colors for each virus}
\usage{
plot_crosstable2(
  df,
  x = "virus",
  y = "domain",
  colors = c(COVID = "#619150", RSV = "#366895", Influenza = "#F1C232"),
  na.rm = TRUE,
  show_missing_x = FALSE,
  show_missing_y = FALSE,
  show_empty_levels = TRUE,
  show_zeros = FALSE,
  show_legend = FALSE,
  legend_title = "No. of studies",
  x_axis_pos = "top",
  x_axis_label_angle = 45,
  x_axis_title = NULL,
  y_axis_title = NULL,
  label_size = 18,
  font_size = 16,
  plot_margin = NULL,
  aspect_ratio = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{x, y}{Categorical variables that will form x and y axis, respectively. Enter as length 1 character vectors}

\item{colors}{Named vector of colors for each virus}

\item{na.rm}{If \code{TRUE}, missing levels are removed from factor variables}

\item{show_missing_x, show_missing_y, show_empty_levels}{If \code{TRUE}, all levels of axis variables will be displayed. If \code{FALSE}, only levels with studies will be displayed in plot}

\item{show_zeros}{If \code{TRUE}, cells with no studies will be displayed as "0". If \code{FALSE}, no 0 is displayed is these cells}

\item{show_legend}{If \code{TRUE}, legend is displayed. Default is \code{FALSE}}

\item{legend_title}{Title for legend}

\item{x_axis_pos}{Position for x axis labels. Default is \code{"top"}}

\item{x_axis_label_angle}{Angle for x axis tick labels}

\item{x_axis_title}{Title for x axis}

\item{y_axis_title}{Title for y axis}

\item{label_size}{Font size of row/column titles}

\item{font_size}{Font size for number of studies}

\item{plot_margin}{Plot margins}

\item{aspect_ratio}{Aspect ratio for plot}

\item{...}{Not used}
}
\value{
ggplot object
}
\description{
Version of \code{plot_crosstable} that uses separate colors for each virus
}
\examples{
plot_crosstable(VIP::ve, x = "virus", y = "vax_product")
plot_crosstable2(VIP::domain, x = "virus", y = "domain")

}
