% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{popover_btn}
\alias{popover_btn}
\title{Popover button}
\usage{
popover_btn(
  ...,
  icon = NULL,
  icon_color = "white",
  color = "#00498F",
  hover_text = "",
  title = "",
  id_popover = NULL,
  id_btn = NULL
)
}
\arguments{
\item{...}{Components to appear in popover upon trigger. Arguments passed to \code{bslib::popover}}

\item{icon}{Icon to display (clicking icon will trigger popover)}

\item{icon_color}{Color of \code{icon}. Default is \code{"white"}}

\item{color}{Color of circle popover button that contains \code{icon} inside. Default is \code{"#00498F"}}

\item{hover_text}{Text to display when hovering over \code{icon}. Default is empty (\code{""})}

\item{title}{Title for popover card. Default is blank (\code{""})}

\item{id_popover}{Popover id}

\item{id_btn}{Button id}
}
\value{
Enter as element in UI
}
\description{
Popover button
}
