% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_importance_scores}
\alias{sort_importance_scores}
\title{#' @keywords internal
permute_columns <- function(x, columns = NULL) {
if (is.null(columns)) {
stop("No columns specified for permutation.")
}
x\link{, columns} <- x\link{sample(nrow(x)), columns}
x
}}
\usage{
sort_importance_scores(x, decreasing)
}
\description{
#' @keywords internal
permute_columns <- function(x, columns = NULL) {
if (is.null(columns)) {
stop("No columns specified for permutation.")
}
x\link{, columns} <- x\link{sample(nrow(x)), columns}
x
}
}
\keyword{internal}
