% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{theme_vip}
\alias{theme_vip}
\title{Custom ggplot2 theme}
\usage{
theme_vip(
  base_size = 14,
  axis_line_thickness = 0.7,
  aspect_ratio = ratio,
  ratio = aspect_ratio,
  y_axis_title_angle = 90,
  y_axis_title_font_size = base_size + 2,
  x_axis_label_angle = 0,
  x_angle = x_axis_label_angle,
  x_axis_label_hjust = 0.5,
  x_axis_label_vjust = 0.5,
  x_axis_label_margin_right = 0,
  x_axis_title_font_size = base_size + 2,
  ...
)
}
\arguments{
\item{base_size}{Size of text in pts. Default is \code{16}}

\item{axis_line_thickness}{Axis and tick thickness in mm. Default is \code{0.7}}

\item{aspect_ratio}{Aspect ratio, y/x. Default is \code{1}}

\item{ratio}{Alias for \code{aspect_ratio}}

\item{y_axis_title_angle}{Angle of text used for y axis title. Default is 90}

\item{x_axis_label_angle}{Angle for x axis tick labels}

\item{x_angle}{Alias for \code{x_axis_label_angle}}

\item{x_axis_label_hjust, x_axis_label_vjust}{Horizontal and vertical justification of x axis text. Default is \code{0.5}}

\item{x_axis_label_margin_right}{Margin to the right of x axis tick labels. Default is \code{0}}

\item{x_axis_title_font_size, y_axis_title_font_size}{Default is \code{base_size + 2}}

\item{...}{Arguments passed to \code{ggplot2::theme}}
}
\value{
theme object
}
\description{
Custom ggplot2 theme
}
