% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Survival of Titanic passengers}
\format{
A data frame with 1309 observations on the following 6 variables:
\itemize{
\item \code{survived} - binary with levels \code{"yes"} for survived and \code{"no"}
otherwise;
\item \code{pclass} - integer giving the corresponding passenger (i.e., ticket)
class with values 1--3;
\item \code{age} - the age in years of the corresponding passenger (with 263
missing values);
\item \code{age} - factor giving the sex of each passenger with levels
\code{"male"} and \code{"female"};
\item \code{sibsp} - integer giving the number of siblings/spouses aboard for each
passenger (ranges from 0--8);
\item \code{parch} - integer giving the number of parents/children aboard for each
passenger (ranges from 0--9).
}
}
\source{
\url{https://hbiostat.org/data/}.
}
\usage{
titanic
}
\description{
A data set containing the survival outcome, passenger class, age, sex, and
the number of family members for a large number of passengers aboard the
ill-fated Titanic.
}
\note{
As mentioned in the column description, \code{age} contains 263 \code{NA}s (or
missing values). For a complete version (or versions) of the data set, see
\link{titanic_mice}.
}
\keyword{datasets}
