% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ve}
\alias{ve}
\title{Vaccine effectiveness data}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
ve
}
\description{
1 row for each combination of patient population, vaccine product, and VE outcome per study
}
\keyword{datasets}
