% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ve_meta}
\alias{ve_meta}
\title{VE meta-analysis results including sensitivity analysis}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
ve_meta
}
\description{
Distinguish core meta-analysis from sensitivity meta-analysis using "sensitivity_criteria" column
}
\keyword{datasets}
