% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ve_meta_raw}
\alias{ve_meta_raw}
\title{Raw data for VE meta-analysis}
\format{
Data frame
\describe{
\item{id_redcap}{RedCAP id}
}
}
\usage{
ve_meta_raw
}
\description{
Raw data for VE meta-analysis
}
\keyword{datasets}
