% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_ace.R
\name{vi_ace}
\alias{vi_ace}
\alias{vi_ace.default}
\alias{vi_ace.formula}
\title{ACE-based variable importance}
\usage{
vi_ace(x, y = NULL, ...)

\method{vi_ace}{default}(x, y = NULL, ...)

\method{vi_ace}{formula}(formula, data, ..., na.action = stats::na.fail)
}
\arguments{
\item{x, formula}{A data frame or a matrix of features, or a formula of the
form \code{response ~ features}.}

\item{y}{A vector of response values}

\item{...}{Additional optional arguments passed onto
\code{\link[acepack]{ace}}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{as.data.frame} to a data frame) containing the variables in the
model. If not found in \code{data}, the variables are taken from
\code{environment(formula)}, typically the environment from which
\code{\link{vi_ace}} was called.}
}
\description{
Compute PDP-based variable importance scores for the predictors in a model.
(This function is meant for internal use only.)
}
\details{
The alternating conditional expectations (ACE) algorithm finds the
transformations of response and features that maximize the proportion of
variation in the response explained by each predictor. The ACE-based variable
importance measure is computed by taking the square of the Pearson
correlation coefficient between each feature and the response.
}
