% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_ice.R
\name{vi_ice}
\alias{vi_ice}
\alias{vi_ice.default}
\title{ICE-Based Variable Importance}
\usage{
vi_ice(object, ...)

\method{vi_ice}{default}(object, feature_names, FUN = NULL,
  var_fun = NULL, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{...}{Additional optional arguments to be passed onto
\code{\link[pdp]{partial}}.}

\item{feature_names}{Character string giving the names of the predictor
variables (i.e., features) of interest.}

\item{FUN}{Deprecated. Use \code{var_fun} instead.}

\item{var_fun}{List with two components, \code{"cat"} and \code{"con"},
containing the functions to use to quantify the variability of the feature
effects (e.g., partial dependence values) for categorical and continuous
features, respectively. If \code{NULL}, the standard deviation is used for
continuous features. For categorical features, the range statistic is used
(i.e., (max - min) / 4). Only used when \code{method = "pdp"} or
\code{method = "ice"}.}
}
\value{
A tidy data frame (i.e., a \code{"tibble"} object) with two columns,
\code{Variable} and \code{Importance}, containing the variable name and its
associated importance score, respectively.
}
\description{
Compute ICE-based variable importance scores for the predictors in a model.
(This function is meant for internal use only.)
}
\details{
Coming soon!
}
