% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vint.R
\name{vint}
\alias{vint}
\title{Interaction Effects}
\usage{
vint(object, feature_names, progress = "none", parallel = FALSE,
  paropts = NULL, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{feature_names}{Character string giving the names of the two features of
interest.}

\item{progress}{Character string giving the name of the progress bar to use
while constructing the interaction statistics. See
\code{\link[plyr]{create_progress_bar}} for details. Default is
\code{"none"}.}

\item{parallel}{Logical indicating whether or not to run \code{partial} in
parallel using a backend provided by the \code{foreach} package. Default is
\code{FALSE}.}

\item{paropts}{List containing additional options to be passed onto
\code{\link[foreach]{foreach}} when \code{parallel = TRUE}.}

\item{...}{Additional optional arguments to be passed onto
\code{\link[pdp]{partial}}.}
}
\description{
Compute the strength of two-way interaction effects. For details, see the
reference below.
}
\details{
Coming soon!
}
\references{
Greenwell, B. M., Boehmke, B. C., and McCarthy, A. J.: A Simple
and Effective Model-Based Variable Importance Measure. arXiv preprint
arXiv:1805.04755 (2018).
}
