% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{vip_shiny}
\alias{vip_shiny}
\title{Run shiny app}
\usage{
vip_shiny(
  data_ve = tryElse(ve),
  data_ae = tryElse(ae),
  data_domain = tryElse(core),
  data_stats = NULL,
  primary_color = "#246A87",
  secondary_color = "#A63B86",
  slider_color = primary_color,
  accordion_fill_color = primary_color,
  accordion_text_color = secondary_color,
  switch_color = primary_color,
  pills = FALSE,
  crosstab_fn = plot_crosstable2,
  clean_data = TRUE
)
}
\arguments{
\item{data_ve, data_ve, data_domain, data_stats}{Data frames for vaccine efficacy, safety, domain, and statistics}

\item{primary_color, secondary_color}{Primary and secondary color for application}

\item{slider_color, switch_color}{Color for sliders and switches}

\item{accordion_fill_color, accordion_text_color}{Color for background and text of accordions}

\item{pills}{If \code{TRUE}, tabs are displayed as pills}

\item{crosstab_fn}{Function used for generating heatmap of study counts}

\item{clean_data}{If \code{TRUE}, only studies using terminology for efficacy and safety outcomes are included in app}
}
\value{
Shiny app run in new window
}
\description{
Run shiny app
}
