% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_v1.R
\name{vip_shiny_v1}
\alias{vip_shiny_v1}
\title{Run shiny app}
\usage{
vip_shiny_v1(
  data = tryElse(df_shiny),
  primary_color = "#246A87",
  secondary_color = "#A63B86",
  slider_color = primary_color,
  accordion_fill_color = primary_color,
  accordion_text_color = secondary_color,
  switch_color = primary_color,
  pills = FALSE,
  crosstab_fn = plot_crosstable2
)
}
\arguments{
\item{data}{Data frame}

\item{primary_color, secondary_color}{Primary and secondary color for application}

\item{slider_color, switch_color}{Color for sliders and switches}

\item{accordion_fill_color, accordion_text_color}{Color for background and text of accordions}

\item{pills}{If \code{TRUE}, tabs are displayed as pills}

\item{crosstab_fn}{Function used for generating heatmap of study counts}
}
\value{
Shiny app run in new window
}
\description{
Run shiny app
}
