% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdVarImp.R
\name{pdVarImp}
\alias{pdVarImp}
\title{Partial Dependence-based Variable Importance}
\usage{
pdVarImp(object, pred.var, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{pred.var}{Character string giving the names of the predictor variables
of interest. For reasons of computation/interpretation, this should include
no more than three variables.}

\item{...}{Additional optional arguments to be passed on to
\code{\link[pdp]{partial}}.}
}
\description{
Compute a variable importance score for the predictor in a model using the
corresponding partial dependence function.
}
