% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest plot}
\usage{
plot_forest(
  df,
  estimate,
  lower,
  upper,
  grouping_var = NULL,
  label_var = "variable",
  ordering_var = NULL,
  dodge_width = 0.7,
  x_axis_title = NULL,
  estimate_label = "regular",
  color_var = grouping_var,
  y_axis_label_size = 10,
  y_axis_label_color = "grey30",
  hjust = 1,
  add_n_bands = NULL,
  n_bands = NULL,
  x_min = NULL,
  x_max = NULL,
  colors = c("#0072B5", "#BC3C29", "#333333"),
  point_shape = "square",
  point_size = 2,
  errorbar_width = 0.25,
  line_thickness = 0.9,
  point_border_thickness = 0.7,
  vert_line_x_position = NULL,
  vert_line_thickness = 0.5,
  vert_line_color = "#333333",
  vert_linetype = "dashed",
  x_label_fn = NULL,
  scale = "regular",
  n_rows = 1,
  scales = "fixed",
  show_stripes = TRUE,
  stripe_color = "#E5E5E5",
  show_legend = FALSE,
  legend_title = NULL,
  base_size = 14,
  ratio = 2.5,
  aspect_ratio = ratio,
  axis_line_thickness = 0.7,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{estimate, lower, upper}{Columns in \code{df} containing effect estimate, lower, and upper bound of confidence interval, respectively}

\item{grouping_var}{Grouping variable along y axis}

\item{label_var}{Variable used to create row labels in plot}

\item{ordering_var}{Variable used to reorder rows in forest plot. If \code{NULL} (default), no reordering}

\item{dodge_width}{Width of dodging. Only relevant when \code{grouping_var} is specified. Default is \code{0.7}}

\item{x_axis_title}{Title for x axis}

\item{estimate_label}{Label for effect estimate}

\item{color_var}{Variable used to determine color of effect estimate/CI bars}

\item{y_axis_label_size}{Font size for y axis labels. Default is \code{10}}

\item{y_axis_label_color}{Font color for y axis labels. Default is \code{"grey30"}}

\item{hjust}{Horizontal adjustment. Default is \code{1}}

\item{add_n_bands}{Number of extra grey/white bands to add above}

\item{n_bands}{Total number of grey/white bands}

\item{x_min, x_max}{Lower and upper limit for x axis values, respectively}

\item{colors}{Colors for effect estimate/CI bars}

\item{point_shape}{Shapes used to display effect estimate. Options: \code{square} (default), \code{circle}, \code{diamond}, \code{triangle}}

\item{point_size}{Size of effect estimate points. Default is \code{2}}

\item{errorbar_width}{Width of error bars. Default is \code{0.25}}

\item{line_thickness}{Line thickness for effect estimate/CI bars. Default is \code{0.9}}

\item{point_border_thickness}{Border thickness for effect estimate points. Default is \code{0.7}}

\item{vert_line_x_position}{x intercept of dashed line. Default is \code{NULL} (no dashed line)}

\item{vert_line_thickness}{Thickness of reference vertical line. Default is \code{0.5}}

\item{vert_line_color}{Color of vertical line. Default is \code{"#333333"}}

\item{vert_linetype}{Linetype for reference vertical line. Default is \code{"dashed"}}

\item{x_label_fn}{Function used to generate x axis labels}

\item{scale}{Scale for effect estimate. Default is \code{"regular"}}

\item{n_rows}{Number of rows. Default is \code{1}}

\item{scales}{Input to \code{scales} argument of \code{facet_wrap} or \code{facet_grid}. Default is \code{"fixed"}}

\item{show_stripes}{If \code{TRUE} (default), alternating white and grey rectangles displayed in background}

\item{stripe_color}{Color for stripes. Default is \code{"#E5E5E5"}}

\item{show_legend}{If \code{FALSE} (default), no legend is displayed}

\item{legend_title}{Title for legend. Only relevant when \code{show_legend = TRUE}}

\item{base_size}{Base font size for plot. Default is \code{14}}

\item{ratio}{Aspect ratio of plot. Default is \code{2.5}}

\item{axis_line_thickness}{Thickness of x axis line and ticks. Default is \code{0.7}}

\item{...}{Arguments passed to \code{theme_vip}}
}
\value{
ggplot object
}
\description{
Forest plot
}
