% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{vi}
\alias{vi}
\alias{vi.default}
\alias{vi.lm}
\title{Variable Importance}
\usage{
vi(object, ...)

\method{vi}{default}(object, pred.var, FUN = NULL, ...)

\method{vi}{lm}(object, use.partial = FALSE, FUN = NULL, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{...}{Additional optional arguments.}

\item{pred.var}{Character string giving the names of the predictor variables
of interest.}

\item{FUN}{Function used to measure the "flatness" of the partial dependence
function. If \code{NULL}, the variance is used (i.e., \code{FUN = var}).}

\item{use.partial}{Logical indicating whether or not to use partial
dependence functions to construct variable importance scores. Default is
\code{FALSE}.}
}
\value{
A tidy data frame (i.e., a \code{"tibble"} object).
}
\description{
Compute variable importance scores for the predictors in a model.
}
