% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{vi}
\alias{vi}
\alias{vi.default}
\alias{vi.earth}
\alias{vi.gbm}
\alias{vi.H2ORegressionModel}
\alias{vi.lm}
\alias{vi.randomForest}
\alias{vi.RandomForest}
\alias{vi.ranger}
\alias{vi.train}
\alias{vi.xgb.Booster}
\title{Variable Importance}
\usage{
vi(object, ...)

\method{vi}{default}(object, pred.var, FUN = NULL, keep.partial = FALSE,
  ...)

\method{vi}{earth}(object, pred.var, partial = FALSE, FUN = NULL,
  keep.partial = FALSE, ...)

\method{vi}{gbm}(object, pred.var, partial = FALSE, FUN = NULL,
  keep.partial = FALSE, ...)

\method{vi}{H2ORegressionModel}(object, pred.var, partial = FALSE,
  FUN = NULL, keep.partial = FALSE, ...)

\method{vi}{lm}(object, pred.var, partial = FALSE, FUN = NULL,
  keep.partial = FALSE, ...)

\method{vi}{randomForest}(object, pred.var, type = 1, partial = FALSE,
  FUN = NULL, keep.partial = FALSE, ...)

\method{vi}{RandomForest}(object, pred.var, auc = FALSE, partial = FALSE,
  FUN = NULL, keep.partial = FALSE, ...)

\method{vi}{ranger}(object, pred.var, type = 1, partial = FALSE,
  FUN = NULL, keep.partial = FALSE, ...)

\method{vi}{train}(object, pred.var, partial = FALSE, FUN = NULL,
  keep.partial = FALSE, ...)

\method{vi}{xgb.Booster}(object, pred.var, type = c("Gain", "Cover",
  "Frequency"), partial = FALSE, FUN = NULL, keep.partial = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{...}{Additional optional arguments.}

\item{pred.var}{Character string giving the names of the predictor variables
of interest.}

\item{FUN}{Function used to measure the variability of the partial dependence
values for continuous predictors. If \code{NULL}, the standard deviation is
used (i.e., \code{FUN = sd}). For factors, the range statistic is used (i.e.,
(max - min) / 4).}

\item{keep.partial}{Logical indicating whether or not to return the computed
partial dependence values for each predictor listed in \code{pred.var} in a
separate attribute called \code{"partial"}. Default is \code{FALSE}.}

\item{partial}{Logical indicating whether or not to use partial dependence
functions to construct variable importance scores. Default is \code{FALSE}.}

\item{type}{Integer specifying the type of variable importance measure to
return for \code{\link[randomForest]{randomForest}} objects. Should be
\code{0} or \code{1}; see \code{\link[randomForest]{importance}} for details.}

\item{auc}{Logical indicating whether or not to compute the AUC-based
variable scores described in Janitza et al. (2012). Only available for
\code{\link[cforest]{party}} objects. See
\code{\link[varimpAUC]{party}} for details. Default is \code{FALSE}.}
}
\value{
A tidy data frame (i.e., a \code{"tibble"} object) with two columns:
\code{Variable} and \code{Importance}.
}
\description{
Compute variable importance scores for the predictors in a model.
}
\examples{
#
# A projection pursuit regression example
#

# Load the sample data
data(mtcars)

# Fit a projection pursuit regression model
mtcars.ppr <- ppr(mpg ~ ., data = mtcars, nterms = 1)

# Compute variable importance scores
vi(mtcars.ppr, pred.var = names(subset(mtcars, select = -mpg)))
}
