% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{vi}
\alias{vi}
\alias{vi.default}
\title{Variable Importance}
\usage{
vi(object, ...)

\method{vi}{default}(object, y, data, method = c("ace", "model", "pdp",
  "ice", "permute"), feature_names, FUN = NULL,
  abbreviate_feature_names = NULL, sort = TRUE, decreasing = TRUE,
  scale = FALSE, rank = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object) or
an object that inherits from class \code{"vi"}. Can also be a a data frame or
a matrix of predictors, or a formula of the form \code{response ~ features}
(only applicable when \code{method = "ace"}).}

\item{...}{Additional optional arguments.}

\item{y}{A vector of response values. Only udes when \code{method = "ace"}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{as.data.frame} to a data frame) containing the variables in the
model. If not found in \code{data}, the variables are taken from
\code{environment(formula)}, typically the environment from which
\code{\link{vi_ace}} was called. Only udes when \code{method = "ace"}.}

\item{method}{Character string specifying the type of variable importance
(VI) to compute. Current options are \code{"ace"}, for ACE-based VI scores
(see \code{\link{vi_ace}} for details); \code{"model"}, for model-based VI
scores (see \code{\link{vi_model}} for details), \code{"pdp"}, for PDP-based
VI scores (see \code{\link{vi_pdp}} for details), \code{"ice"}, for ICE-based
VI scores (see \code{\link{vi_ice}} for details), and \code{"permute"}, for
permutation-based VI scores (see \code{\link{vi_permute}} for details). The
default depends on the class of \code{object}. For more details on the
PDP/ICE-based methods, see the reference below.}

\item{feature_names}{Character string giving the names of the predictor
variables (i.e., features) of interest.}

\item{FUN}{List with two components, \code{"cat"} and \code{"con"},
containing the functions to use for categorical and continuous features,
respectively. If \code{NULL}, the standard deviation is used for continuous
features. For categorical features, the range statistic is used (i.e.,
(max - min) / 4).}

\item{abbreviate_feature_names}{Integer specifying the length at which to
abbreviate feature names. Default is \code{NULL} which results in no abbreviation
(i.e., the full name of each feature will be printed).}

\item{sort}{Logical indicating whether or not to order the sort the variable
importance scores. Default is \code{TRUE}.}

\item{decreasing}{Logical indicating whether or not the variable importance
scores should be sorted in descending (\code{TRUE}) or ascending
(\code{FALSE}) order of importance. Default is \code{TRUE}.}

\item{scale}{Logical indicating whether or not to scale the variable
importance scores so that the largest is 100. Default is \code{FALSE}.}

\item{rank}{Logical indicating whether or not to rank the variable
importance scores (i.e., convert to integer ranks). Default is \code{FALSE}.
Potentially useful when comparing variable importance scores across different
models using different methods.}
}
\value{
A tidy data frame (i.e., a \code{"tibble"} object) with two columns:
\code{Variable} and \code{Importance}. For \code{"lm"/"glm"}-like objects, an
additional column, called \code{Sign}, is also included which includes the
sign (i.e., POS/NEG) of the original coefficient.
}
\description{
Compute variable importance scores for the predictors in a model.
}
\examples{
#
# A projection pursuit regression example
#

# Load the sample data
data(mtcars)

# Fit a projection pursuit regression model
mtcars.ppr <- ppr(mpg ~ ., data = mtcars, nterms = 1)

# Compute variable importance scores
vi(mtcars.ppr, method = "ice")

# Plot variable importance scores
vip(mtcars.ppr, method = "ice")
}
\references{
Greenwell, B. M., Boehmke, B. C., and McCarthy, A. J. A Simple
and Effective Model-Based Variable Importance Measure. arXiv preprint
arXiv:1805.04755 (2018).
}
