% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_permute.R
\name{vi_permute}
\alias{vi_permute}
\title{Permutation-Based Variable Importance}
\usage{
vi_permute(object, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{feature_names}{Character string giving the names of the predictor
variables (i.e., features) of interest.}

\item{train}{Non-optional data frame containing the original training data.}

\item{pfun}{Non-optional prediction function that requires two arguments:
\code{object} and \code{newdata}. If specified, then the function must return
a vector of predictions (i.e., not a matrix or data frame). (In the future,
this argument may become optional.)}

\item{obs}{Non-optional vector containing the original (i.e., training)
response values.}

\item{metric}{Non-optional function for computing model performance (e.g.,
RMSE for regression or accuracy for binary classification). This function
requires two arguments, \code{pred} (for predicted values) and \code{obs}
(for observed values), and should return a single, numeric value.}

\item{Non-optional}{character string specifying whether this is supervized
regression (\code{"regression"}) or classification (\code{"classification"}).}
}
\value{
A tidy data frame (i.e., a \code{"tibble"} object) with two columns:
\code{Variable} and \code{Importance}. For \code{"glm"}-like object, an
additional column, called \code{Sign}, is also included which gives the sign
(i.e., POS/NEG) of the original coefficient.
}
\description{
Compute permutation-based variable importance scores for the predictors in a
model. (This function is meant for internal use only.)
}
\details{
Coming soon!
}
