% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vint.R
\name{vint}
\alias{vint}
\title{Interaction Effects}
\usage{
vint(object, pred.var, progress = "none", parallel = FALSE,
  paropts = NULL, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{pred.var}{Character string giving the names of the predictor variables
of interest. Should be of length two.}

\item{progress}{Character string giving the name of the progress bar to use
while constructing the interaction statistics. See
\code{\link[plyr]{create_progress_bar}} for details. Default is
\code{"none"}.}

\item{parallel}{Logical indicating whether or not to run \code{partial} in
parallel using a backend provided by the \code{foreach} package. Default is
\code{FALSE}.}

\item{paropts}{List containing additional options to be passed onto
\code{\link[foreach]{foreach}} when \code{parallel = TRUE}.}

\item{...}{Additional optional arguments to be passed onto
\code{\link[pdp]{partial}}.}
}
\description{
Compute the strength of interaction effects.
}
\details{
Coming soon!
}
\note{
Coming soon!
}
