% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{vip}
\alias{vip}
\alias{vip.default}
\title{Variable Importance Plots}
\usage{
vip(object, ...)

\method{vip}{default}(object, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{...}{Additional optional arguments to be passed onto \code{\link{vi}}.}
}
\description{
Plot variable importance scores for the predictors in a model.
}
\examples{
\dontrun{
mtcars.lm <- lm(mpg ~ ., data = mtcars)
vip(mtcars.lm) + theme_light()
}
}
