% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{vip}
\alias{vip}
\alias{vip.default}
\title{Variable Importance Plots}
\usage{
vip(object, ...)

\method{vip}{default}(object, alpha = 1, color = "grey35",
  fill = "grey35", ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object).}

\item{...}{Additional optional arguments to be passed onto \code{\link{vi}}.}

\item{alpha}{Numeric value between 0 and 1 giving the trasparency of the
bars.}

\item{color}{Character string specifying the color to use for the borders of
the bars. Could also be a function, such as
\code{\link[grDevices]{heat.colors}}. Default is \code{"grey35"}.}

\item{fill}{Character string specifying the color to use to fill the bars.
Could also be a function, such as \code{\link[grDevices]{heat.colors}}.
Default is \code{"grey35"}.}
}
\description{
Plot variable importance scores for the predictors in a model.
}
\examples{
\dontrun{
mtcars.lm <- lm(mpg ~ ., data = mtcars)
vip(mtcars.lm) + theme_light()
}
}
