% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{vip}
\alias{vip}
\alias{vip.default}
\alias{vip.model_fit}
\title{Variable Importance Plots}
\usage{
vip(object, ...)

\method{vip}{default}(
  object,
  num_features = 10L,
  type = c("barplot", "dotplot", "boxplot"),
  mapping = NULL,
  aesthetics = list(),
  horizontal = TRUE,
  all_permutations = FALSE,
  jitter = FALSE,
  include_type = FALSE,
  ...
)

\method{vip}{model_fit}(object, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., a \code{"randomForest"} object) or
an object that inherits from class \code{"vi"}.}

\item{...}{Additional optional arguments to be passed on to \code{\link{vi}}.}

\item{num_features}{Integer specifying the number of variable importance
scores to plot. Default is \code{10}.}

\item{type}{Character string specifying which type of plot to construct.
Current options are \code{"barplot"} (the default), \code{"dotplot"} (for
Cleveland dot plots), or \code{"boxplot"}; these may be abbreviated as well
(e.g., \code{"bar"}, \code{"dot"}, and \code{"box"}, respectively).
Specifying \code{type = "boxplot"} is only useful for the permutation-based
importance method with \code{nsim > 1} and \code{keep = TRUE}; see
\code{\link{vi_permute}} for details.}

\item{mapping}{Set of aesthetic mappings created by \code{link[ggplot2]{aes}}
or \code{link[ggplot2]{aes_}}. See example usage below.}

\item{aesthetics}{List specifying additional arguments passed on to
\code{link[ggplot2]{layer}}. These are often aesthetics, used to set an
aesthetic to a fixed value, like \code{colour = "red"} or \code{size = 3}.
See example usage below.}

\item{horizontal}{Logical indicating whether or not to plot the importance
scores on the x-axis (\code{TRUE}). Default is \code{TRUE}.}

\item{all_permutations}{Logical indicating whether or not to plot all
permutation scores along with the average. Default is \code{FALSE}. (Only
used for permutation scores when \code{nsim > 1}.)}

\item{jitter}{Logical indicating whether or not to jitter the raw permutation
scores. Default is \code{FALSE}. (Only used when
\code{all_permutations = TRUE}.)}

\item{include_type}{Logical indicating whether or not to include the type of
variable importance computed in the axis label. Default is \code{FALSE}.}
}
\description{
Plot variable importance scores for the predictors in a model.
}
\examples{
#
# A projection pursuit regression example
#

# Load the sample data
data(mtcars)

# Fit a projection pursuit regression model
model <- ppr(mpg ~ ., data = mtcars, nterms = 1)

# Construct variable importance plot
vip(model, method = "ice")

# Better yet, store the variable importance scores and then plot
vi_scores <- vi(model, method = "ice")
vip(vi_scores, type = "dotplot", horiz = FALSE)
vip(vi_scores, type = "dotplot", horiz = FALSE, aesthetics = list(size = 3))

# The `\%T>\\%` operator is imported for convenience; see ?magrittr::`\%T>\%`
# for details
vi_scores <- model \%>\%
  vi(method = "ice") \%T>\%
  {print(vip(.))}
vi_scores

# Permutation scores (barplot w/ raw values and jittering)
vip(model, method = "permute", train = mtcars, target = "mpg", nsim = 10,
    metric = "rmse", aesthetics = list(color = "grey50", fill = "grey50"),
    all_permutations = TRUE, jitter = TRUE)

# Permutation scores (boxplot)
vip(model, method = "permute", train = mtcars, target = "mpg", nsim = 10,
    metric = "rmse", type = "boxplot")

# Permutation scores (boxplot colored by feature)
library(ggplot2)  # for `aes_string()` function
vip(model, method = "permute", train = mtcars, target = "mpg", nsim = 10,
    metric = "rmse", type = "boxplot", all_permutations = TRUE,
    mapping = aes_string(fill = "Variable"),
    aesthetics = list(color = "grey35", size = 0.8))
}
